requirements.txt
alembic==1.7.4
blinker==1.4
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.5
MarkupSafe==2.0.1
PyJWT==2.3.0
python-dotenv==0.19.1
SQLAlchemy==1.4.26
Werkzeug==2.0.2
WTForms==2.3.3
